+++ lib/dashboard/index.js
@@ -6,8 +6,9 @@ const path = require('path');
 const pug = require('pug');
 
 // Packages
-const express = require('express');
 const clone = require('clone');
+const debounce = require('lodash.debounce');
+const express = require('express');
 
 // Ours
 const configHelper = require('../config');
@@ -20,6 +21,10 @@ const PUBLIC_PATH = path.join(__dirname, 'public');
 const WEB_APP_MANIFEST_PATH = path.join(__dirname, 'manifest.json');
 const SERVICE_WORKER_PATH = path.join(__dirname, 'service-worker.js');
 let cachedDashboard = null;
+let workspaces;
+const recacheDashboard = debounce(() => {
+	cachedDashboard = renderDashboard();
+}, 100);
 
 log.trace('Adding Express routes');
 app.use('/bower_components', express.static(path.resolve(__dirname, '../../bower_components')));
@@ -90,11 +95,10 @@ app.get('/panel/:bundle/*', ncgUtils.authCheck, (req, res, next) => {
 });
 
 // Initialize cache
-cacheDashboard();
+recacheDashboard();
 
-// When a panel is re-parsed, re-render and cache the dashboard
-bundles.on('bundleChanged', cacheDashboard);
-
+// When a bundle changes, re-cache the dashboard.
+bundles.on('bundleChanged', recacheDashboard);
 module.exports = app;
 
 function renderDashboard() {
@@ -112,13 +116,35 @@ function renderDashboard() {
 		}),
 		publicConfig: configHelper.filteredConfig,
 		privateConfig: configHelper.config,
-		workspaces: [{
-			name: 'Main Workspace',
-			route: ''
-		}]
+		workspaces: parseWorkspaces()
 	});
 }
 
-function cacheDashboard() {
-	cachedDashboard = renderDashboard();
+function parseWorkspaces() {
+	const workspaceNames = new Set();
+	bundles.all().forEach(bundle => {
+		bundle.dashboard.panels.forEach(panel => {
+			if (panel.workspace) {
+				workspaceNames.add(panel.workspace);
+			}
+		});
+	});
+
+	const workspaces = [{
+		name: 'default',
+		route: ''
+	}];
+
+	workspaceNames.forEach(name => {
+		if (name === 'default') {
+			return;
+		}
+
+		workspaces.push({
+			name,
+			route: name
+		});
+	});
+
+	return workspaces;
 }
