+++ lib/dashboard/public/elements/ncg-dashboard.js
@@ -24,15 +24,12 @@
 				type: Array,
 				value() {
 					const workspaces = window.__renderData__.workspaces;
-					if (workspaces[0].name === 'Main Workspace' && workspaces.length === 1) {
-						workspaces[0].name = 'Workspace';
-					}
+
+					// Give the default workspace a more meaningful display name
+					workspaces[0].name = workspaces.length === 1 ? 'Workspace' : 'Main Workspace';
+
 					return workspaces;
 				}
-			},
-			panels: {
-				type: Array,
-				computed: '_computePanels(bundles)'
 			},
 			dialogs: {
 				type: Array,
@@ -70,20 +67,8 @@
 			}
 		},
 
-		_routeChanged(route) {
-			// This is a hack to fix packery when the viewport size is changed
-			// when the workspace is not visible.
-			if (route.path === '') {
-				this.debounce('fixPackery', this._fixPackery, 10);
-			}
-
+		_routeChanged() {
 			this._fixTabs();
-		},
-
-		_fixPackery() {
-			if (this.route.path === '') {
-				this.applyPackery();
-			}
 		},
 
 		_fixTabs() {
@@ -97,18 +82,6 @@
 					}
 				});
 			}
-		},
-
-		_computePanels(bundles) {
-			const panels = [];
-			bundles.forEach(bundle => {
-				bundle.dashboard.panels.forEach(panel => {
-					if (!panel.dialog) {
-						panels.push(panel);
-					}
-				});
-			});
-			return panels;
 		},
 
 		_computeDialogs(bundles) {
@@ -194,44 +167,28 @@
 				});
 			});
 
-			window.addEventListener('load', () => {
-				this.$.loadingSpinner.active = false;
-				setTimeout(() => {
-					this.$.panels.style.opacity = 1;
-					this.$.panels.style.transform = 'translateY(0)';
-					this.$.panels.style.pointerEvents = 'auto';
-				}, 750);
-
-				/*
-				 * Packery
-				 */
-				window.imagesLoaded(document, () => {
-					// Init Packery
-					this.initPackery();
-					this.startObservingPanelMutations();
-
-					// Once all the panel iFrames are loaded, this func (from the iframe-resize bower dep)
-					// automagically fixes the height of all the iframes.
-					window.iFrameResize({
-						log: false,
-						resizeFrom: 'child',
-						heightCalculationMethod: 'documentElementOffset',
-						resizedCallback(data) {
-							data.iframe.dispatchEvent(new CustomEvent('iframe-resized'));
-						}
-					});
-
-					// Sometimes, we just need to know when a dang click event occurred. No matter where it happened.
-					// This adds a `panelClick` event to all panels.
-					/* eslint-disable no-loop-func */
-					const panels = document.querySelectorAll('ncg-dashboard-panel iframe');
-					for (let i = 0; i < panels.length; i++) {
-						panels[i].contentDocument.addEventListener('click', e => {
-							document.dispatchEvent(new CustomEvent('panelClick', e.target));
-						});
+			window.imagesLoaded(document, () => {
+				// Once all the panel iFrames are loaded, this func (from the iframe-resize bower dep)
+				// automagically fixes the height of all the iframes.
+				window.iFrameResize({
+					log: false,
+					resizeFrom: 'child',
+					heightCalculationMethod: 'documentElementOffset',
+					resizedCallback(data) {
+						data.iframe.dispatchEvent(new CustomEvent('iframe-resized'));
 					}
-					/* eslint-enable no-loop-func */
 				});
+
+				// Sometimes, we just need to know when a dang click event occurred. No matter where it happened.
+				// This adds a `panelClick` event to all panels.
+				/* eslint-disable no-loop-func */
+				const panels = document.querySelectorAll('ncg-dashboard-panel iframe');
+				for (let i = 0; i < panels.length; i++) {
+					panels[i].contentDocument.addEventListener('click', e => {
+						document.dispatchEvent(new CustomEvent('panelClick', e.target));
+					});
+				}
+				/* eslint-enable no-loop-func */
 			});
 		},
 
@@ -255,153 +212,6 @@
 			this.querySelector('app-header-layout').$.contentContainer.style.boxSizing = 'border-box';
 		},
 
-		startObservingPanelMutations() {
-			// Packery causes attributes changes, so before applying it
-			// we disconnect then reconnect our observer to avoid infinite loops
-			try {
-				if (this._panelMutationObserver) {
-					return;
-				}
-
-				// Create a MutationObserver which will watch for changes to the DOM and re-apply masonry
-				this._panelMutationObserver = new MutationObserver(() => {
-					this.debounce('handleMutation', this._debouncedMutationHandler, 150);
-				});
-
-				// Define what element should be observed by the observer
-				// and what types of mutations trigger the callback
-				this._panelMutationObserver.observe(this.$.panels, {
-					subtree: true,
-					attributes: true,
-					childList: true,
-					characterData: true,
-					attributeOldValue: false,
-					characterDataOldValue: false
-				});
-			} catch (e) {
-				console.warn('MutationObserver not supported, dashboard panels may be less responsive to DOM changes');
-			}
-		},
-
-		initPackery() {
-			const packery = new Packery(this.$.panels, {
-				itemSelector: 'ncg-dashboard-panel',
-				columnWidth: 128,
-				gutter: 16,
-				isInitLayout: false,
-				containerStyle: {position: 'relative'}
-			});
-			this._packery = packery;
-
-			// Initial sort
-			const sortOrder = []; // global variable for saving order, used later
-			let storedSortOrder = localStorage.getItem('panelSortingOrder');
-			if (storedSortOrder) {
-				storedSortOrder = JSON.parse(storedSortOrder);
-
-				// Create a hash of items
-				const itemsByFullName = {};
-				const allPanels = [];
-				let panelName;
-				let bundleName;
-				let i;
-				let len;
-				for (i = 0, len = packery.items.length; i < len; i++) {
-					const item = packery.items[i];
-					panelName = item.element.getAttribute('panel');
-					bundleName = item.element.getAttribute('bundle');
-					const fullName = bundleName + '.' + panelName;
-					allPanels[i] = fullName;
-					itemsByFullName[fullName] = item;
-				}
-
-				// Merge the saved panel array with our currently loaded panels, remove dupes
-				const allPanelsOrdered = arrayUnique(storedSortOrder.concat(allPanels));
-
-				// Remove panels that no longer exist
-				const removededOld = allPanelsOrdered.filter(val => {
-					return allPanels.indexOf(val) !== -1;
-				});
-
-				// Overwrite packery item order
-				len = removededOld.length;
-				for (i = 0; i < len; i++) {
-					panelName = removededOld[i];
-					packery.items[i] = itemsByFullName[panelName];
-				}
-			}
-
-			// Manually trigger initial layout
-			packery.layout();
-			this._packeryInitialized = true;
-
-			const panelsList = this.$.panels.querySelectorAll('ncg-dashboard-panel');
-			panelsList.forEach(itemElem => {
-				// Make element draggable with Draggabilly
-				const draggie = new Draggabilly(itemElem, {handle: '#dragHandle'});
-
-				// Bind Draggabilly events to Packery
-				packery.bindDraggabillyEvents(draggie);
-			});
-
-			// Currently, there is no built-in option to force dragged elements to gravitate to their
-			// nearest neighbour in a specific direction. This will reset their locations 100ms after a drag
-			// causing them to gravitate.
-			packery.on('dragItemPositioned', () => {
-				setTimeout(() => {
-					packery.layout();
-				}, 100);
-			});
-
-			packery.on('layoutComplete', orderItems);
-			packery.on('dragItemPositioned', orderItems);
-
-			function orderItems() {
-				const itemElems = packery.getItemElements();
-
-				// Reset / empty order array
-				sortOrder.length = 0;
-
-				for (let i = 0; i < itemElems.length; i++) {
-					sortOrder[i] = itemElems[i].getAttribute('bundle') + '.' + itemElems[i].getAttribute('panel');
-				}
-
-				// Save ordering
-				localStorage.setItem('panelSortingOrder', JSON.stringify(sortOrder));
-			}
-		},
-
-		applyPackery() {
-			this.debounce('applyPackery', () => {
-				if (this._packeryInitialized) {
-					this._packery.layout();
-				}
-			}, 10);
-		},
-
-		shiftPackery() {
-			// Called when anything in #panels receives a click or tap event.
-			// There's probably a lot of room to optimize this and not call
-			// this routine for every single click,
-			// but this was just the easiest way to ensure that mutations
-			// caused by clicks are caught, because those mutations might
-			// have changed the vertical size of the panel.
-
-			this.debounce('shiftPackery', () => {
-				if (this._packeryInitialized) {
-					// See http://packery.metafizzy.co/methods.html#shiftlayout for more details
-					this._packery.shiftLayout();
-				}
-			}, 10);
-		},
-
-		_handlePanelCollapseTransition(e) {
-			// Whenever a panel finishes transitioning, shiftPackery.
-			if (e.detail.value === false) {
-				this.shiftPackery();
-			}
-		},
-
 		logout() {
 			window.location.href = '/logout';
 		},
@@ -410,28 +220,10 @@
 			this.$.drawer.close();
 		},
 
-		_debouncedMutationHandler() {
-			this._panelMutationObserver.disconnect();
-			this.shiftPackery();
-			this.startObservingPanelMutations();
-		},
-
 		_equal(a, b) {
 			return a === b;
 		}
 	});
-
-	function arrayUnique(array) {
-		const a = array.concat();
-		for (let i = 0; i < a.length; ++i) {
-			for (let j = i + 1; j < a.length; ++j) {
-				if (a[i] === a[j]) {
-					a.splice(j--, 1);
-				}
-			}
-		}
-		return a;
-	}
 
 	function getImageDataURI(url, cb) {
 		let data;
