+++ package.json
@@ -10,7 +10,7 @@
   "bugs": "https://github.com/nodecg/nodecg/issues",
   "license": "MIT",
   "dependencies": {
-    "@nodecg/bundle-manager": "^0.5.3",
+    "@nodecg/bundle-manager": "^0.5.4",
     "@nodecg/logger": "^0.2.0",
     "body-parser": "^1.14.1",
     "cheerio": "^0.22.0",
@@ -27,11 +27,11 @@
     "httpolyglot": "^0.1.1",
     "is-my-json-valid": "^2.12.3",
     "json-schema-defaults": "^0.3.0",
+    "lodash.debounce": "^4.0.8",
     "minimist": "^1.2.0",
     "multer": "^1.1.0",
     "nedb": "^1.2.1",
     "node-localstorage": "^1.3.0",
-    "nodecg-bundle-parser": "^0.3.2",
     "object-path": "^0.11.1",
     "parse-authors": "^0.2.2",
     "passport": "^0.3.2",
